added plt-racket PPA
This commit probably works, but there are some changes with git-fetch
and such.

+++ setup.sh
@@ -5,29 +5,32 @@ mkdir -p $HOME/projects-vanilla
 mkdir -p $HOME/projects
 
 DESKTOP="${DESKTOP:=0}"
+DEPLOY="${DEPLOY:=0}"
+LATEST_VIM="${LATEST_VIM:=0}"
 
 sudo add-apt-repository -y ppa:synapse-core/testing
 sudo apt-add-repository -y ppa:ubuntu-mozilla-daily/firefox-aurora
+sudo apt-add-repository -y ppa:plt/racket
 sudo apt-add-repository -y ppa:pi-rho/dev
-sudo apt-get update
-sudo apt-get install -y git mercurial build-essential vim-gtk emacs tmux ncdu \
+sudo apt-get -y update
+sudo apt-get -y install git mercurial build-essential emacs tmux ncdu \
     lftp curl elinks cloc autossh feh htop rsync rlwrap virtualbox postgresql \
     postgresql-client python-dev exuberant-ctags acpi mosh openssh-server ranger \
-    tig tree 
+    tig tree plt-racket
 
 if [ $DESKTOP -eq 1 ]; then
-    sudo apt-get install -y synapse pinta mupdf mplayer vlc vlc-nox firefox flashplugin-installer
+    sudo apt-get -y install synapse pinta mupdf mplayer vlc vlc-nox firefox flashplugin-installer
 fi 
 
 if [ $LATEST_VIM -eq 1]; then
-    sudo apt-get install -y vim-gtk
+    sudo apt-get -y install vim-gtk
 fi
 
 (
     cd $HOME/projects-vanilla/
 
     ( #fasd
-        git clone https://github.com/clvv/fasd.git
+        git clone https://github.com/clvv/fasd.git; (cd fasd; git fetch; git reset --hard origin/master)
         cd fasd
         PREFIX=$HOME make install
     )
@@ -41,13 +44,13 @@ fi
     )
 
     ( #java, clojure, maven, leiningen
-        git clone https://github.com/technomancy/leiningen.git
-        git clone https://github.com/flexiondotorg/oab-java6.git
+        git clone https://github.com/technomancy/leiningen.git; ( cd leiningen; git fetch; git reset --hard origin/master)
+        git clone https://github.com/flexiondotorg/oab-java6.git; ( cd oab-java6.git; git fetch; git reset --hard origin/master)
     )
 
     ( #zsh
-        git clone https://github.com/zsh-users/zsh.git
-        git clone https://github.com/zsh-users/antigen.git
+        git clone https://github.com/zsh-users/zsh.git; ( cd zsh; git fetch ; git reset --hard origin/master)
+        git clone https://github.com/zsh-users/antigen.git; ( cd antigen; git fetch ; git reset --hard origin/master)
         cd zsh
         git checkout tags/zsh-5.0.5
     )
@@ -57,10 +60,10 @@ fi
         git clone https://github.com/creationix/nvm.git $HOME/.nvm
         source $HOME/.nvm/nvm.sh
         # TODO
-        nvm install 0.10
-        nvm alias default 0.10
-        nvm use 0.10
-        #npm install -g http-server underscore-cli bower supervisor
+        #nvm install 0.10
+        #nvm alias default 0.10
+        #nvm use 0.10
+        #npm install -g http-server underscore-cli bower supervisor pm2
     )
 
     ( #TODO throwaway ruby, perl, python
@@ -70,24 +73,26 @@ fi
     ( #ack
         # should be https
         # instructions from http://beyondgrep.com/install/
-        curl http://beyondgrep.com/ack-2.12-single-file > ~/bin/ack && chmod 0755 !#:3 
+        curl http://beyondgrep.com/ack-2.12-single-file > ~/bin/ack && chmod 0755 ~/bin/ack
     )
 
 )
 
-( #dotfiles
-    git clone https://github.com/shurane/dotfiles.git $HOME/dotfiles
-    cd $HOME/dotfiles
-    for elem in ".bashrc" ".pathrc" ".loadrc" ".bash_aliases" ".vimrc" ".tmux.conf" ".zshrc" ".ackrc" ".inputrc" ".gitconfig"; do
-        rm $HOME/$elem
-        ln -s $(readlink -f $elem) $HOME/
-    done
-)
-
 ( #vim
-    git clone https://github.com/Shougo/neobundle.vim ~/.vim/bundle/neobundle.vim
+    git clone https://github.com/Shougo/neobundle.vim $HOME/.vim/bundle/neobundle.vim; (cd $HOME/.vim/bundle/neobundle.vim; git fetch; git reset --hard origin/master)
     #git clone https://github.com/gmarik/vundle.git $HOME/.vim/bundle/vundle
     vim +NeoBundleInstall! +qall
+)
+
+( #dotfiles
+    if [ $DEPLOY -eq 1 ]; then
+        git clone https://github.com/shurane/dotfiles.git $HOME/dotfiles
+        cd $HOME/dotfiles
+        for elem in ".bashrc" ".pathrc" ".loadrc" ".bash_aliases" ".vimrc" ".tmux.conf" ".zshrc" ".ackrc" ".inputrc" ".gitconfig"; do
+            rm $HOME/$elem
+            ln -s $(readlink -f $elem) $HOME/
+        done
+    fi
 )
 
 source $HOME/.bashrc
